+++ src/gl/planes.h
@@ -1,0 +1,8 @@
+#include "gl.h"
+
+#ifndef __PLANE_H_
+#define __PLANE_H_
+
+void gl4es_glClipPlanef(GLenum plane, const GLfloat *equation);
+
+#endif